+++ faucet-sensitive.json
@@ -1,0 +1,4 @@
+{
+    "FAUCET_ADDRESS": "FAUCET_ADDRESS_SHOULDNT_BE_AT_STATIC",
+    "FAUCET_PRIVATE_KEY": "FAUCET_PRIVATE_SHOULDNT_BE_AT_STATIC"
+}
