+++ package.json
@@ -21,7 +21,7 @@
     "ethereumjs-tx": "1.3.6",
     "ganache-cli": "^6.7.0",
     "ganache-core": "^2.8.0",
-    "next": "9.1.1",
+    "next": "9.3.2",
     "next-images": "^1.2.0",
     "node-schedule": "^1.3.2",
     "react": "16.10.2",
