+++ prod-config.json
@@ -3,8 +3,7 @@
   "API_URL": "NO_API",
   "NEW_CAPTCHA_URL": "https://rust-captcha.herokuapp.com/new/easy/5/998",
   "SOLVE_CAPTCHA_URL": "https://rust-captcha.herokuapp.com/solution/",
-  "FAUCET_ADDRESS": "NO_ADDRESS",
-  "FAUCET_PRIVATE_KEY": "NO_PRIV",
+  "CAPTCHA_API_URL": "https://cors-anywhere.herokuapp.com/https://rust-captcha.herokuapp.com",
   "GAS_PRICE": 60000000,
   "GAS_LIMIT": 100000,
   "VALUE_TO_DISPENSE": 0.05,
